+++ main.js
@@ -1128,6 +1128,23 @@ function serializeSelected(event, method = 'LR') {
     }, 250);
 }
 
+function deserializeSelected(event, method = 'LR') {
+    files.forEach(f => f.meta.checked ? f.source?.stop() : '');
+    document.getElementById('loadingText').textContent = 'Processing';
+    document.body.classList.add('loading');
+    setTimeout(() => {
+        const selected = files.filter(
+          f => f.meta.checked && f.buffer.numberOfChannels === 1);
+        selected.forEach((f, idx) => {
+            editor.deserialize(event, f, false, method);
+            if (idx === selected.length - 1) {
+                document.body.classList.remove('loading');
+            }
+        });
+        renderList();
+    }, 250);
+}
+
 function pitchUpSelected(event) {
     files.forEach(f => f.meta.checked ? f.source?.stop() : '');
     document.getElementById('loadingText').textContent = 'Processing';
@@ -5127,6 +5144,7 @@ window.digichain = {
     stretchSelected,
     shortenNameSelected,
     serializeSelected,
+    deserializeSelected,
     condenseSelected,
     nudgeCrossingsSelected,
     padWithZeroSelected,